Refactor task base classes to retrieve generic actions
