+++ src/main/groovy/wooga/gradle/unity/tasks/AbstractBatchModeTask.groovy
@@ -2,13 +2,33 @@ package wooga.gradle.unity.tasks
 
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Optional
+import org.gradle.api.tasks.TaskAction
+import org.gradle.process.ExecResult
+import wooga.gradle.unity.batchMode.BaseBatchModeSpec
+import wooga.gradle.unity.batchMode.BatchModeAction
 import wooga.gradle.unity.batchMode.BatchModeSpec
 import wooga.gradle.unity.batchMode.BuildTarget
 
 abstract class AbstractBatchModeTask extends AbstractUnityTask implements BatchModeSpec {
 
+    @Delegate(excludeTypes = [ExecuteExclude.class], interfaces = false)
+    protected BatchModeAction batchModeAction
+
+    private ExecResult batchModeResult
+
     AbstractBatchModeTask(Class taskType) {
         super(taskType)
+        this.batchModeAction = retrieveBatchModeActionFactory().create()
+    }
+
+    @Override
+    BaseBatchModeSpec retrieveAction() {
+        return batchModeAction
+    }
+
+    @TaskAction
+    protected void exec() {
+        batchModeResult = batchModeAction.execute()
     }
 
     /**
@@ -46,5 +66,13 @@ abstract class AbstractBatchModeTask extends AbstractUnityTask implements BatchM
     @Override
     Boolean getNoGraphics() {
         batchModeAction.noGraphics
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Input
+    boolean isIgnoreExitValue() {
+        return batchModeAction.isIgnoreExitValue()
     }
 }
